% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easing.R
\name{anime_easing_spring}
\alias{anime_easing_spring}
\title{Specify a spring physics easing}
\usage{
anime_easing_spring(mass = 1, stiffness = 100, damping = 10, velocity = 0)
}
\arguments{
\item{mass}{Numeric. Object mass.}

\item{stiffness}{Numeric. Spring stiffness.}

\item{damping}{Numeric. Damping coefficient.}

\item{velocity}{Numeric. Initial velocity.}
}
\value{
A character string of the form
\code{"spring(mass,stiffness,damping,velocity)"}.
}
\description{
Constructs the Anime.js v4 \code{spring(mass, stiffness, damping, velocity)}
easing string.
}
\examples{
anime_easing_spring()
anime_easing_spring(mass = 2, stiffness = 80, damping = 5, velocity = 1)

}
