% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{anime_on}
\alias{anime_on}
\title{Attach a JavaScript callback to an animation event}
\usage{
anime_on(x, event, callback)
}
\arguments{
\item{x}{An \code{anime_timeline} or \code{anime_animation} object.}

\item{event}{One of
"onBegin", "onBeforeUpdate", "onUpdate", "onRender", "onLoop", "onPause", "onComplete",
matching the Anime.js v4 callback API.}

\item{callback}{Character scalar. Name of the global JS function to invoke.}
}
\value{
The modified object.
}
\description{
The callback must be the name of a globally scoped JavaScript function
already present on the page, for example one injected via
\code{htmltools::tags$script()}. At render time the JavaScript binding resolves
the name to \code{window[callback]} and attaches it to the corresponding
Anime.js callback.
}
\examples{
if (interactive() && rlang::is_installed("htmltools")) {
  svg_src <- '<svg viewBox="0 0 200 100" xmlns="http://www.w3.org/2000/svg">
    <circle class="circle" cx="100" cy="50" r="20" fill="#4e79a7"/>
  </svg>'

  widget <- anime_timeline(duration = 800) |>
    anime_add(selector = ".circle", props = list(opacity = c(0, 1))) |>
    anime_on("onComplete", "handleAnimationDone") |>
    anime_render(svg = svg_src)

  callback_js <- htmltools::tags$script(
    "function handleAnimationDone() {
      console.log('Animation complete.');
    }"
  )

  htmltools::browsable(htmltools::tagList(callback_js, widget))
}

}
