% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{anime_target_class}
\alias{anime_target_class}
\title{Target elements by CSS class}
\usage{
anime_target_class(cls)
}
\arguments{
\item{cls}{Character scalar. Class name without a leading dot.}
}
\value{
A CSS selector string of the form \code{".<cls>"}.
}
\description{
Target elements by CSS class
}
\examples{
anime_target_class("circle")

}
