% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{anime_target_id}
\alias{anime_target_id}
\title{Target SVG or HTML elements by a data-animejs-id attribute}
\usage{
anime_target_id(id)
}
\arguments{
\item{id}{Character scalar. Value of the \code{data-animejs-id} attribute.}
}
\value{
A CSS selector string of the form \code{"[data-animejs-id='<id>']"}.
}
\description{
The primary mechanism for targeting individual elements annotated by
\code{svg_annotate()} (in \code{gganime}) or by hand.
}
\examples{
anime_target_id("c1")

}
