% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{anime_timeline}
\alias{anime_timeline}
\title{Initialise an Anime.js timeline}
\usage{
anime_timeline(duration = 1000, ease = anime_easing(), delay = 0, loop = FALSE)
}
\arguments{
\item{duration}{Default duration in milliseconds for all segments.}

\item{ease}{Default easing for all segments.}

\item{delay}{Default delay in milliseconds between segments.}

\item{loop}{Logical or integer. \code{FALSE} for no looping, \code{TRUE} for infinite
looping, or a positive integer for a fixed number of iterations.}
}
\value{
An \code{anime_timeline} object.
}
\description{
Initialise an Anime.js timeline
}
\examples{
anime_timeline(duration = 800, ease = anime_easing())

}
