% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{animejs-shiny}
\alias{animejs-shiny}
\alias{animejsOutput}
\alias{renderAnimejs}
\title{Shiny bindings for animejs}
\usage{
animejsOutput(outputId, width = "100\%", height = "400px")

renderAnimejs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"}, \code{"400px"},
\code{"auto"}) or a number, which will be coerced to a string and have \code{"px"}
appended.}

\item{expr}{An expression that generates an animejs widget, typically a
call to \code{\link[=anime_render]{anime_render()}}.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This is
useful if you want to save an expression in a variable.}
}
\value{
\code{animejsOutput()} returns a Shiny output function that can be used
in a UI definition; \code{renderAnimejs()} returns a Shiny render function
that can be assigned to an output slot.
}
\description{
Output and render functions for using animejs widgets within Shiny
applications and interactive R Markdown documents.
}
\examples{
if (interactive() && rlang::is_installed("shiny")) {
  library(shiny)

  svg_src <- '<svg viewBox="0 0 200 100" xmlns="http://www.w3.org/2000/svg">
    <circle class="dot" cx="100" cy="50" r="20" fill="#4e79a7"/>
  </svg>'

  ui <- fluidPage(
    sliderInput("duration", "Duration (ms)", 200, 2000, 800),
    animejsOutput("anim", height = "200px")
  )

  server <- function(input, output, session) {
    output$anim <- renderAnimejs({
      anime_animate(
        selector = anime_target_class("dot"),
        props = list(opacity = anime_from_to(0, 1)),
        duration = input$duration
      ) |>
        anime_render(svg = svg_src)
    })
  }

  shinyApp(ui, server)
}

}
