% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_duration}
\alias{validate_duration}
\title{Validate a duration or delay value}
\usage{
validate_duration(x, arg = "duration")
}
\arguments{
\item{x}{Value to validate.}

\item{arg}{Argument name used in error messages.}
}
\value{
\code{x} invisibly if valid.
}
\description{
Validate a duration or delay value
}
\keyword{internal}
