% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{anime_timeline}
\alias{anime_timeline}
\title{Initialise an Anime.js timeline}
\usage{
anime_timeline(
  duration = 1000,
  ease = anime_easing(),
  loop = FALSE,
  delay = 0,
  direction = "normal"
)
}
\arguments{
\item{duration}{Default duration in milliseconds for all segments.}

\item{ease}{Default easing for all segments.}

\item{loop}{Logical. Whether the timeline loops indefinitely.}

\item{delay}{Default delay in milliseconds between segments.}

\item{direction}{One of \code{"normal"}, \code{"reverse"}, \code{"alternate"}.}
}
\value{
An \code{anime_timeline} object.
}
\description{
Initialise an Anime.js timeline
}
